change recursive call to loop
