+++ server/src/gpt/gpt.ts
@@ -1,151 +1,95 @@
-import axios from 'axios'
-import FormData from 'form-data'
-import * as fs from 'fs';
-import { ChatCompletionResponse, GPTFunction, Message } from './types'
-import _ from 'lodash'
-import { FunctionRegistry } from './functionRegistry'
-
-const baseContext: Message[] = [
-  {
-    role: "system",
-    content: "Du er en AI assistent. Svarene dine vil være korte, presise og høflige"
-  },
-  {
-    role: "system",
-    content: "Gi en hyggelig velkomstmelding til din første bruker"
-  }
-]
-
-export const transcribeAudio = async (openAIKey: string, filePath: string, model: string): Promise<any> => {
-  // Create a new FormData instance
-  const form = new FormData()
-
-  // Add file and model to the form
-  form.append('file', fs.createReadStream(filePath))
-  form.append('model', model)
-  form.append('language', 'no')
-
-  // Make the request to the OpenAI API
-  try {
-    
-    const response = await axios.post('https://api.openai.com/v1/audio/transcriptions', form, {
-      headers: {
-        ...form.getHeaders(),
-        'Authorization': `Bearer ${openAIKey}`,
-      },
-    })
-
-    return response.data.text
-  } catch (error) {
-    console.error(`Fetch to OpenAI API failed: ${error}`)
-    throw error
-  }
-}
+import axios from "axios"
+import { FunctionRegistry } from "./functionRegistry"
+import { ChatCompletionResponse, GPTFunction, Message } from "./types"
+import _ from "lodash"
 
 const chatCompletion = async (
-  openAIKey: string, 
-  messages: Message[],
-  functionRegistry: FunctionRegistry
-  ): Promise<ChatCompletionResponse> => {
-  const response = await axios.post(
-    'https://api.openai.com/v1/chat/completions',
-    {
-      model: "gpt-4-0613",
-      messages: messages,
-      functions: functionRegistry.getOPENAIReadyFunctions()
-    },
-    {
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${openAIKey}`
+    openAIKey: string, 
+    messages: Message[],
+    functionRegistry: FunctionRegistry
+    ): Promise<ChatCompletionResponse> => {
+    const response = await axios.post(
+      'https://api.openai.com/v1/chat/completions',
+      {
+        model: "gpt-4-0613",
+        messages: messages,
+        functions: functionRegistry.getOPENAIReadyFunctions()
       },
+      {
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${openAIKey}`
+        },
+      }
+    )
+    if(response.status !== 200) {
+      throw new Error(`Failed when communicating with OPENAI API ${response}`)
     }
-  )
-  if(response.status !== 200) {
-    throw new Error(`Failed when communicating with OPENAI API ${response}`)
+    return response.data
   }
-  return response.data
+  
+
+export const callGptLoop = async (
+    openAIKey: string,
+    originalMessages: Message[], 
+    functionRegistry: FunctionRegistry
+):Promise<{messages: Message[], tokensUsed:number}> => {
+    // Call GPT
+    let gptResponse = await chatCompletion(openAIKey, originalMessages, functionRegistry)
+    let functionCall = gptResponse.choices[0].message.function_call
+    const newMessages: Message[] = _.cloneDeep(originalMessages)
+
+    // While gpt suggests functions, keep calling them
+    while(functionCall) {
+        // Extract arguments from the suggested function call
+        const args = functionCall.arguments
+        // Retrieve the actual function from the function registry
+        const requestedFunction = functionRegistry.getRequestedFuntion(functionCall.name)
+        if(!requestedFunction) {
+            originalMessages.push({
+              role: 'system',
+              content: `The function does not exist, dont try calling it again!'}`
+            })
+          } else {
+            const functionMessages = await callFunction(requestedFunction, args)
+            originalMessages.push(...functionMessages)
+        }
+        // Call gpt with the function result
+        gptResponse = await chatCompletion(openAIKey, originalMessages, functionRegistry)
+        functionCall = gptResponse.choices[0].message.function_call
+    }
+
+    // Gpt didnt suggest any new functions, we return the response to client
+    newMessages.push({
+        role: gptResponse.choices[0].message.role,
+        content: gptResponse.choices[0].message.content
+    })
+    return {
+        messages: newMessages,
+        tokensUsed: gptResponse.usage.total_tokens
+    }
 }
 
-export const handleGptResponse = async (
-  openAIKey: string, 
-  response: ChatCompletionResponse, 
-  messages: Message[], 
-  functionRegistry: FunctionRegistry
-  ): Promise<number> => {
-    const functionCall = response.choices[0].message.function_call
-
-    if(!functionCall) { // GPT does not suggest any function
-      messages.push({
-        role: response.choices[0].message.role,
-        content: response.choices[0].message.content
-      })
-      return response.usage.total_tokens
-    }
-    // Extract arguments from the suggested function call
-    const args = functionCall.arguments
-
-    // Retrieve the actual function from the function registry
-    const requestedFunction = functionRegistry.getRequestedFuntion(functionCall.name)
-    
-    if(!requestedFunction) {
+// Calls the function suggested by GPT with args suggested by GPT
+const callFunction = async (gptFunction: GPTFunction, args: string): Promise<Message[]> => {
+    const messages: Message[] = []
+    messages.push({
+      role: 'system',
+      content: `You called this function ${gptFunction.name} with these arguments ${args}`
+    })
+    try {
+        console.log(`Calling func suggested by gpt ${gptFunction.name} with args ${args}`)
+        const result = await gptFunction.func(JSON.parse(args))
+        messages.push({
+            role: 'system',
+            content: `Result of the functioncall: ${JSON.stringify(result)}`
+        })
+    } catch (error: unknown) {
+      console.log(`Functioncall failed ${(error as Error).message}`)
       messages.push({
         role: 'system',
-        content: `The function does not exist, dont try calling it again!'}`
+        content: `Dont try this function ${gptFunction.name} with these arguments ${args} again!`
       })
-    } else {
-      const relatedMessages = await callFunction(requestedFunction, args)
-      messages.push(...relatedMessages)
     }
-    const gptResponse = await chatCompletion(openAIKey, messages, functionRegistry)
-    await handleGptResponse(openAIKey, gptResponse, messages, functionRegistry)
-    return response.usage.total_tokens
-}
-
-export const handleMessages = async (
-  openAIKey: string,
-  messages: Message[], 
-  functionRegistry: FunctionRegistry
-  ): Promise<number> => {
-    const gptResponse = await chatCompletion(openAIKey, messages, functionRegistry)
-    return await handleGptResponse(openAIKey, gptResponse, messages, functionRegistry)
-}
-
-export const getDefaultMessages = (): Message[] => {
-  return _.cloneDeep(baseContext)
-} 
-
-const callFunction = async (gptFunction: GPTFunction, args: string): Promise<Message[]> => {
-  const messages: Message[] = []
-  messages.push({
-    role: 'system',
-    content: `You called this function ${gptFunction.name} with these arguments ${args}`
-  })
-  try {
-    const result = await gptFunction.func(JSON.parse(args))
-    messages.push({
-      role: 'system',
-      content: `Result of the functioncall: ${JSON.stringify(result)}`
-    })
-  } catch (error: unknown) {
-    console.log(`Functioncall failed ${(error as Error).message}`)
-    messages.push({
-      role: 'system',
-      content: `Dont try this function ${gptFunction.name} with these arguments ${args} again!`
-    })
+    return messages
   }
-  return messages
-}
-
-
-
-
-// Til API
-const messages=[
-      {"role": "system", "content": "You are a helpful assistant."},
-      {"role": "user", "content": "Who won the world series in 2020?"}
-  ]
-// Fra API
-const answer = [
-  {"role": "assistant", "content": "The Los Angeles Dodgers won the World Series in 2020."},
-]
